Tested

+++ src/com/models/UserModel.java
@@ -65,7 +65,34 @@ public class UserModel {
 	public void setLon(Double lon) {
 		this.lon = lon;
 	}
-
+	
+	/**
+	 * 
+	 * @param userEmail
+	 * @return userID
+	 * @throws SQLException
+	 */
+	static int getuserID (String userEmail) throws SQLException{
+		
+		Connection conn = DBConnection.getActiveConnection();
+		String sql = ("Select `id` from users where `email` = '"+userEmail+"'");
+		PreparedStatement stmt;
+		stmt = conn.prepareStatement(sql);
+		ResultSet rs = stmt.executeQuery();
+		int userID = 0;
+		if(rs.next()){
+			userID = rs.getInt("id");
+		}
+		return userID;
+	}
+	
+	/**
+	 * 
+	 * @param name
+	 * @param email
+	 * @param pass
+	 * @return user
+	 */
 	public static UserModel addNewUser(String name, String email, String pass) {
 		try {
 			Connection conn = DBConnection.getActiveConnection();
@@ -96,9 +123,13 @@ public class UserModel {
 		}
 		return null;
 	}
-
 	
-	
+	/**
+	 * 
+	 * @param email
+	 * @param pass
+	 * @return user
+	 */
 	public static UserModel login(String email, String pass) {
 		try {
 			Connection conn = DBConnection.getActiveConnection();
@@ -125,7 +156,14 @@ public class UserModel {
 		}
 		return null;
 	}
-
+	
+	/**
+	 * 
+	 * @param id
+	 * @param lat
+	 * @param lon
+	 * @return true
+	 */
 	public static boolean updateUserPosition(Integer id, Double lat, Double lon) {
 		try{
 			Connection conn = DBConnection.getActiveConnection();
@@ -143,6 +181,12 @@ public class UserModel {
 		return false;
 	}
 	
+	/**
+	 * 
+	 * @param id
+	 * @return user
+	 * @throws SQLException
+	 */
 	public static UserModel LastPosition(Integer id) throws SQLException {
 		try{
 			Connection conn = DBConnection.getActiveConnection();
@@ -165,33 +209,23 @@ public class UserModel {
 		return null;	
 	}
 	
-	public static Boolean follow(String email1, String email2) throws SQLException {
+	/**
+	 * 
+	 * @param userEmail
+	 * @param followingUserEmail
+	 * @return true or false
+	 * @throws SQLException
+	 */
+	public static Boolean follow(String userEmail, String followingUserEmail) throws SQLException {
 		try {
 			Connection conn = DBConnection.getActiveConnection();
-			
-			String sql1 = ("Select `id` from users where `email` = ?");
-			String sql2 = ("Select `id` from users where `email` = ?");
 			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql1);
-			stmt.setString(1, email1);
-			ResultSet rs1 = stmt.executeQuery();
-			UserModel user1 = new UserModel();
-			UserModel user2 = new UserModel();
-			if(rs1.next())
-			{
-				user1.id = rs1.getInt("id");
-			}
-			stmt = conn.prepareStatement(sql2);
-			stmt.setString(1, email2);
-			ResultSet rs2 = stmt.executeQuery();
-			if(rs2.next())
-			{
-				user2.id = rs2.getInt("id");
-			}
-			String sql3 = ("Insert into followers (`id1`, `id2`) VALUES  (?,?)");
+
+			int userID = getuserID(userEmail);
+			int followingID = getuserID(followingUserEmail);
+
+			String sql3 = ("Insert into followers (`id1`, `id2`) VALUES  ('"+userID+"','"+followingID+"')");
 			stmt = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
-			stmt.setInt(1, user1.id);
-			stmt.setInt(2, user2.id);
 			stmt.executeUpdate();
 			return true;		
 		}catch(SQLException e){
@@ -199,34 +233,25 @@ public class UserModel {
 		}
 		return false;
 	}
-	public static Boolean unfollow(String email1, String email2)
+	
+	
+	/**
+	 * 
+	 * @param userEmail
+	 * @param followingUserEmail
+	 * @return true or false
+	 */
+	public static Boolean unfollow(String userEmail, String followingUserEmail)
 	{
 		try {
 			Connection conn = DBConnection.getActiveConnection();
-			
-			String sql1 = ("Select `id` from users where `email` = ?");
-			String sql2 = ("Select `id` from users where `email` = ?");
 			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql1);
-			stmt.setString(1, email1);
-			ResultSet rs1 = stmt.executeQuery();
-			UserModel user1 = new UserModel();
-			UserModel user2 = new UserModel();
-			if(rs1.next())
-			{
-				user1.id = rs1.getInt("id");
-			}
-			stmt = conn.prepareStatement(sql2);
-			stmt.setString(1, email2);
-			ResultSet rs2 = stmt.executeQuery();
-			if(rs2.next())
-			{
-				user2.id = rs2.getInt("id");
-			}
-			String sql3 = ("Delete from followers where `id1` = ? and `id2` = ?");
+
+			int userID = getuserID(userEmail);
+			int followingID = getuserID(followingUserEmail);
+
+			String sql3 = ("Delete from followers where `id1` = '"+userID+"' and `id2` = '"+followingID+"'");
 			stmt = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
-			stmt.setInt(1, user1.id);
-			stmt.setInt(2, user2.id);
 			stmt.executeUpdate();
 			return true;		
 		}catch(SQLException e){
@@ -235,7 +260,13 @@ public class UserModel {
 		return false;
 	}
 
-	public static ArrayList<Integer> followers(int id) throws SQLException
+	/**
+	 * 
+	 * @param id
+	 * @return list of following
+	 * @throws SQLException
+	 */
+	public static ArrayList<Integer> following(int id) throws SQLException
 	{
 		Connection conn = DBConnection.getActiveConnection();
 		String sql1 = ("Select `id2` from followers where `id1` = ?");
@@ -290,7 +321,16 @@ public class UserModel {
 		
 	}
 	
-	public static Boolean addNewPlace(String name, String description, Double lat, Double lon) throws SQLException
+	/**
+	 * 
+	 * @param placeName
+	 * @param description
+	 * @param lat
+	 * @param lon
+	 * @return true or false
+	 * @throws SQLException
+	 */
+	public static Boolean addNewPlace(String placeName, String description, Double lat, Double lon) throws SQLException
 	{
 		try 
 		{
@@ -298,7 +338,7 @@ public class UserModel {
 			String sql = "Insert into places (`name`,`description`,`lat`, `long`) VALUES  (?,?,?,?)";
 			PreparedStatement stmt;
 			stmt = conn.prepareStatement(sql);
-			stmt.setString(1, name);
+			stmt.setString(1, placeName);
 			stmt.setString(2, description);
 			stmt.setDouble(3, lat);
 			stmt.setDouble(4, lon);
@@ -310,19 +350,17 @@ public class UserModel {
 		return false;
 	}
 	
-
-	public static Boolean savePlace (String email , String placeName){
+	/**
+	 * 
+	 * @param userEmail
+	 * @param placeName
+	 * @return true or false
+	 */
+	public static Boolean savePlace (String userEmail , String placeName){
 		try{
+
 			Connection conn = DBConnection.getActiveConnection();
-			String sql = ("Select `id` from users where `email` = ?");
-			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql);
-			stmt.setString(1, email);
-			ResultSet rs = stmt.executeQuery();
-			int uid=0;
-			if(rs.next()){
-				uid = rs.getInt("id");
-			}
+			int userID = getuserID(userEmail);
 			
 			String sql2 = ("Select `id` from places where `name` = '"+placeName+"'");
 			PreparedStatement stmt2;
@@ -336,7 +374,7 @@ public class UserModel {
 			String sql3 = ("insert into savedplaces(`user_ID` , `place_ID`) values (? ,?)");
 			PreparedStatement stmt3;
 			stmt3 = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
-			stmt3.setInt(1, uid);
+			stmt3.setInt(1, userID);
 			stmt3.setInt(2, placeID);
 			stmt3.executeUpdate();
 			
@@ -349,18 +387,17 @@ public class UserModel {
 		return false;
 	}
 	
-	
-	public static Boolean checkIn (String email , String placeName ){
+	/**
+	 * 
+	 * @param userEmail
+	 * @param placeName
+	 * @return true or false
+	 */
+	public static Boolean checkIn (String userEmail , String placeName ){
 		try{
 			Connection conn = DBConnection.getActiveConnection();
-			String sql = ("Select `id` from users where `email` = '"+email+"'");
-			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql);
-			ResultSet rs = stmt.executeQuery();
-			int uid=0;
-			if(rs.next()){
-				uid = rs.getInt("id");
-			}
+
+			int userID = getuserID(userEmail);
 			
 			String sql2 = ("Select `id` from places where `name` = '"+placeName+"'");
 			PreparedStatement stmt2;
@@ -371,7 +408,7 @@ public class UserModel {
 				placeID = rs2.getInt("id");
 			}
 			
-			String sql3 = ("insert into checkin(`userID` , `placeID`) values ('"+uid+"' ,'"+placeID+"')");
+			String sql3 = ("insert into checkin(`userID` , `placeID`) values ('"+userID+"' ,'"+placeID+"')");
 			PreparedStatement stmt3;
 			stmt3 = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
 			stmt3.executeUpdate();
@@ -385,39 +422,27 @@ public class UserModel {
 		return false;
 	}
 	
-	public static Boolean like (String email , int checkinID){
+	/**
+	 * 
+	 * @param userEmail
+	 * @param checkinID
+	 * @return true or false
+	 */
+	public static Boolean like (String userEmail , int checkinID){
 		
 		try{
 			Connection conn = DBConnection.getActiveConnection();
-			String sql = ("Select `id` from users where `email` = '"+email+"'");
 			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql);
-			//stmt.setString(1, email);
-			ResultSet rs = stmt.executeQuery();
-			int uid=0;
-			if(rs.next()){
-				uid = rs.getInt("id");
-			}
-			/*
-			String sql2 = ("Select `id` from places where `name` = '"+checkinID+"'");
-			PreparedStatement stmt2;
-			stmt2 = conn.prepareStatement(sql2);
-			ResultSet rs2 = stmt2.executeQuery();
-			int checkid =0;
-			if(rs2.next()){
-				checkid = rs2.getInt("checkinID");
-			}
-			*/
+
+			int userID = getuserID(userEmail);
 			
-			String sql3 = ("insert into likes(`userID` , `checkinID`) values ('"+uid+"' ,'"+checkinID+"')");
+			String sql3 = ("insert into likes(`userID` , `checkinID`) values ('"+userID+"' ,'"+checkinID+"')");
 			PreparedStatement stmt3;
 			stmt3 = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
-//			stmt3.setInt(1, uid);
-//			stmt3.setInt(2, placeID);
 			stmt3.executeUpdate();
 			
 			UserModel m = new UserModel ();
-			m.getnotification(uid, checkinID, "Like");
+			m.notify(userID, checkinID, "Like");
 			return true;
 		}
 		
@@ -427,27 +452,26 @@ public class UserModel {
 		return false;
 	}
 	
-	
-	public static Boolean comment (String email , int checkinID , String desc){
+	/**
+	 * 
+	 * @param userEmail
+	 * @param checkinID
+	 * @param description
+	 * @return true or false
+	 */
+	public static Boolean comment (String userEmail , int checkinID , String description){
 		try{
 			Connection conn = DBConnection.getActiveConnection();
-			String sql = ("Select `id` from users where `email` = '"+email+"'");
-			PreparedStatement stmt;
-			stmt = conn.prepareStatement(sql);
-			//stmt.setString(1, email);
-			ResultSet rs = stmt.executeQuery();
-			int uid=0;
-			if(rs.next()){
-				uid = rs.getInt("id");
-			}
 			
-			String sql3 = ("insert into comment(`user-ID` , `checkin-ID` , `description`) values ('"+uid+"' ,'"+checkinID+"' , '"+desc+"')");
+			int userID = getuserID(userEmail);
+			
+			String sql3 = ("insert into comment(`user-ID` , `checkin-ID` , `description`) values ('"+userID+"' ,'"+checkinID+"' , '"+description+"')");
 			PreparedStatement stmt3;
 			stmt3 = conn.prepareStatement(sql3, Statement.RETURN_GENERATED_KEYS);
 			stmt3.executeUpdate();
 			
 			UserModel m = new UserModel ();
-			m.getnotification(uid, checkinID, "comment");
+			m.notify(userID, checkinID, "comment");
 			
 			return true;
 		}
@@ -457,13 +481,18 @@ public class UserModel {
 		}
 		return false ;
 	}
-
 	
-	
-	public  Boolean getnotification (int userid , int checkinID , String type){
+	/**
+	 * 
+	 * @param userID
+	 * @param checkinID
+	 * @param type
+	 * @return true or false
+	 */
+	public  Boolean notify (int userID , int checkinID , String type){
 		try {
 			Connection conn = DBConnection.getActiveConnection();
-			String sql = ("insert into notification(`userID` , `checkinID` , `type`) values ('"+userid+"' ,'"+checkinID+"' , '"+type+"')");
+			String sql = ("insert into notification(`userID` , `checkinID` , `type`) values ('"+userID+"' ,'"+checkinID+"' , '"+type+"')");
 			PreparedStatement stmt3;
 			stmt3 = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
 			stmt3.executeUpdate();
@@ -475,6 +504,32 @@ public class UserModel {
 			e.printStackTrace();
 		}
 		return false ;
+	}
+	
+	/**
+	 * 
+	 * @param userEmail
+	 * @return list of notifications
+	 * @throws SQLException
+	 */
+	public static ArrayList<Integer> getAllNotification(String userEmail)throws SQLException {
+
+		Connection conn = DBConnection.getActiveConnection();
+
+		int userID = getuserID(userEmail);
+
+		ArrayList<Integer> checkInIds = new ArrayList<Integer>();
+
+		String sql1 = ("select `checkinID` from checkin where `userID` ='"+ userID + "' ");
+		PreparedStatement stmt1;
+		stmt1 = conn.prepareStatement(sql1);
+		ResultSet rs1 = stmt1.executeQuery();
+		int cid = 0;
+		while (rs1.next()) {
+			checkInIds.add(rs1.getInt("checkinID"));
+		}
+		
+		return checkInIds;
 	}
 	
 }
